Producer/requirements.txt
kafka-python
requests
datetime